requirements.txt
alembic
fastapi
httpx
pydantic
pydantic-settings
pytest
pytest-asyncio
pytest-cov
python-dotenv
uvicorn
